+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 
 * Add select ionic radii from CRC Handbook of Chemistry and Physics, 82nd Ed
 * Update ``merge`` module to use multiprocessing
+* Increase efficiency of clash-checking algorithm in ``ldrs_helper.py``
 
 v0.7.2 (2023-07-27)
 ------------------------------------------------------------